Stage 4 implementation: add the new /route command
Add the new /route command to the CommandParser. Add it to the list of valid
commands as well as the switch statement.

Add comments to the switch block to identify the parts of the command list.

Add an empty method for the route command in MetroLine. Implementation will be next.