Add error checking.
Add error checking and error messages throughout the program. It's not
strictly necessary to pass the stages, but it's a good habit to be in to make
the program more robust. Also, the messages make the program more
user-friendly as well.

Add a new method - isValidLineName() - to Controller to avoid duplication of
code throughout the switch block.